requirements.txt
absl-py==0.15.0
aiohttp==3.7.4.post0
apache-beam==2.32.0
applicationinsights==0.11.10
asgiref==3.4.1
asn1crypto==1.4.0
astor==0.8.1
astroid==2.8.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
avro-python3==1.9.2.1
backcall==0.2.0
backports.csv==1.0.7
backports.entry-points-selectable==1.1.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cli-helpers==0.2.3
click==8.0.3
cloud-sql-python-connector @ file:///home/rhugalf/cloud-sql-python-connector
configobj==5.0.6
crcmod==1.7
cryptography==3.4.8
decorator==5.1.0
dill==0.3.1.1
distlib==0.3.3
Django==3.2.9
docopt==0.6.2
enum34==1.1.10
fastavro==1.4.5
fasteners==0.16.3
filelock==3.3.2
flake8==4.0.1
Flask==2.0.2
Flask-PyMongo==2.3.0
fsspec==2021.11.0
future==0.18.2
gast==0.5.2
gce-v2v==0.0.0
gcsfs==2021.11.0
google-api-core==1.31.3
google-api-python-client==2.29.0
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-bigquery==2.30.1
google-cloud-bigtable==1.7.0
google-cloud-core==1.7.2
google-cloud-datastore==1.15.3
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-logging==2.7.0
google-cloud-pubsub==1.7.0
google-cloud-recommendations-ai==0.2.0
google-cloud-spanner==1.19.1
google-cloud-storage==1.42.3
google-cloud-translate==3.6.1
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.1
grpcio-gcp==0.2.2
h5py==3.5.0
hdfs==2.6.0
httplib2==0.19.1
humanize==3.12.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipython==7.29.0
isort==5.10.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
libcst==0.3.21
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mssql-cli==1.0.0
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
orjson==3.6.3
packaging==21.2
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pg8000==1.21.2
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.22
proto-plus==1.19.7
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.8.2
ptyprocess==0.7.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycurl==7.43.0.2
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.2.0
Pygments==2.10.0
PyGObject==3.30.4
pylint==2.11.1
pyls-mypy==0.1.8
pymongo==3.12.1
PyMySQL==0.9.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
pysql-beam==0.2
python-apt==1.8.4.3
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rope==0.21.0
rsa==4.7.2
scramp==1.4.1
six==1.16.0
snowballstemmer==2.1.0
sqlparse==0.4.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminaltables==3.1.0
toml==0.10.2
traitlets==5.1.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
typing-inspect==0.7.1
ujson==4.2.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
yapf==0.31.0
yarl==1.6.3
zipp==3.6.0
